back/requirement.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
Click==7.0
constantly==15.1.0
coverage==4.5.1
cryptography==2.3.1
daphne==2.2.2
Django==2.1.2
django-cors-headers=2.4.0
dynaconf==1.1.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
pip==18.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-box==3.2.2
python-dotenv==0.9.1
pytz==2018.6
service-identity==17.0.0
setuptools==40.5.0
six==1.11.0
toml==0.10.0
Twisted==18.9.0
txaio==18.8.1
wheel==0.32.2
zope.interface==4.6.0
djangorestframework=3.9.0
